adding each methods file
