+++ .rubocop.yml
@@ -8,8 +8,6 @@ AllCops:
 
 Layout/LineLength:
   Max: 120
-Metrics/ModuleLength:
-  Max: 200
 Metrics/MethodLength:
   Max: 20
 Metrics/AbcSize:
